refactor: 1/2 include the common source-code also into the consumer package.
~~Problem: We need the common package with python 3.7 as the minimum
python version for the provisioning listener.~~

Problem: If the consumer package depends on the common package, we have to
publish both packages to enable external consumer development.
It would be much easier for external people if the only depend on one
package.

We need the source code from the common package in the consumer package
with python 3.11 as the minimum python version.

Solution:
Include the common source code in both the common and consumer packages.
To archieve this, the common package needs to be inside the directory of
the consumer package. (Poetry limitation)

Alternative: Remove the "common" package and just install the
"consumer" package in the server components like udm-listener,
disparcher...
Since the udm-listener is now UCS 5.2 -> python 3.11.
This is now possible.