fix(push_config_to_appcenter): the upload script needs to handle both apps
this version of the script uploads the appcenter files from both app
directories (provisioning-service and provisioning-service-backend).

the upload of the backend app scripts is disabled by default. Simply
uncomment the last line of the script if you also need to update the
backend app scripts.

Issue: univention/dev/projects/provisioning#91