chore(renaming): Rename "provisioning-stack" to "provisioning-service"
The old Debian package univention-provisioning-stack-listener is now replaced
by univention-provisioning-service-listener. The new Debian package was
uploaded to the Provider Portal to make it available during app installations
without the need to release it or to switch to the internal mirror. The old
package still exists for the already published but deprecated version of this
app.

Issue: univention/dev/projects/provisioning#91